updated insertIntoContentDB to check for existing documents
In the if-else, a check is performed to find a existing document that matches the title/seasonNumber/episodeNumber for tv shows/movies. If a document is returned, then the existing document is deleted, and the new data from the request is saved in a new document. Otherwise, no documents are deleted, and the new data is saved in a new document.