made modifications; replaced deleteOne with updateOne
Replaced deleteOne query with updateOne to maintain the same ObjectId. If a record is found with the same title/seasonNumber/episodeNumber, then the document is updated with the new data sent in req.body and status: "success" is returned if the update executed successfully. Else, a new document is created.